added BPP as output parameter
added the BPP=Theta*d0/4 to the calculations. For us this is always an important value, knowing that it is easily to calculate from M2. 
I don't know if I have found all poisitions, where to add the BPP, as I could only test one type of analysis with my data.